backend/requirements.txt
fastapi==0.104.1
daphne==4.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.7.7
pillow==10.1.0
httpx==0.25.2
aiofiles==23.2.1

